requirements.txt
mcp[cli]>=1.13.0
requests>=2.32.0
python-dotenv>=1.0.0